Seperated processing from graphite reporting and logging/error cleanup
