+++ src/lib/report_graphite.js
@@ -1,30 +1,16 @@
-var rc_util = require('rippled-network-crawler/src/lib/utility.js');
-var _ = require('lodash');
 var moment = require('moment');
 
-module.exports = function(crawl, graphiteClient) {
-  var data = JSON.parse(crawl.data);
-  var metrics = {
-    crawler: {
-      ippCount: rc_util.getIpps(data).length,
-      publicKeyCount: Object.keys(rc_util.getRippleds(data)).length,
-      connectionsCount: Object.keys(rc_util.getLinks(data)).length,
-      rippleds: {}
-    }
-  };
-  var rippleds = rc_util.getRippledsC(data);
-  _.each(Object.keys(rippleds), function (rippled) {
-    metrics.crawler.rippleds[rippled] = {
-      connectionsCount: rippleds[rippled].in + rippleds[rippled].out,
-      up: 1
-    };
-  });
+// Currently doesn't work. Structure will be different, so it needs to be
+//  rewritten anyways
+
+module.exports = function(metrics, graphiteClient) {
+  // note: crawl not defined
   graphiteClient.write(metrics, moment(crawl.start_at).valueOf(), function(error) {
     if (error) {
-      console.error(error);
+      throw new Error(error);
     } else {
       console.log('Wrote crawl %d \t at %s \t to graphite',
-            crawl.id, moment().format());
+        crawl.id, moment().format());
     }
   });
 };
