+++ src/live.js
@@ -1,7 +1,8 @@
 'use strict';
 var rc_util = require('rippled-network-crawler/src/lib/utility.js');
 var _ = require('lodash');
-var writeToGraphite = require('./lib/report_graphite');
+var reportGraphite = require('./lib/report_graphite');
+var processCrawl = require('./lib/process_crawl');
 var sqs_u = require('./lib/sqs_util');
 var moment = require('moment');
 var Promise = require('bluebird');
@@ -11,10 +12,6 @@ module.exports = function(max, timeout, queueUrl, dbUrl, graphiteUrl) {
   var graphiteClient = graphite.createClient(graphiteUrl);
   var logsql = false;
   var count = 0;
-
-  function processCrawl(crawl) {
-    writeToGraphite(crawl, graphiteClient)
-  }
 
   function processMessage(workerID) {
     sqs_u.getMessage(queueUrl)
@@ -26,25 +23,25 @@ module.exports = function(max, timeout, queueUrl, dbUrl, graphiteUrl) {
         .getRowById(dbUrl, id, logsql)
         .then(function(crawl) {
           if (!crawl) {
-            console.error('No crawls with id %s', id);
-            count -= 1;
-            return -1;
+            throw new Error('No crawls with id %s', id)
           }
 
-          processCrawl(crawl);
-          count -= 1;
+          var crawlMetrics = processCrawl(crawl);
+          // TODO write to db
+          // TODO report to graphite
 
-        }).then(function() {
-          sqs_u.deleteMessage(queueUrl, receiptHandle).then(function() {
-            return 0;
-          });
         })
-        .catch(console.error);
-
-      } else {
-        console.error('No new messages \t at %s', moment().format());
+        .then(function() {  // terminate on success
+          sqs_u.deleteMessage(queueUrl, receiptHandle);
+          count -= 1;
+        })
+        .catch(function(error) { // terminate on message processing error
+          count -= 1;
+          console.error(error);
+        });
+      } else { // terminate on message absence error
         count -= 1;
-        return -1;
+        console.error('No new messages \t at %s', moment().format());
       }
     });
   };
